refactor : 공통 exception(from dev)을 사용하도록 수정
