--- Server/main07/src/main/java/com/codestates/main07/clubBoard/exception/BusinessLogicException.java
@@ -1,11 +1,0 @@
-package com.codestates.main07.clubBoard.exception;
-
-
-public class BusinessLogicException extends RuntimeException {
-    private final ExceptionCode exceptionCode;
-
-    public BusinessLogicException(ExceptionCode exceptionCode) {
-        super(exceptionCode.getMessage());
-        this.exceptionCode = exceptionCode;
-    }
-}
