--- Server/main07/src/main/java/com/codestates/main07/clubBoard/exception/ExceptionCode.java
@@ -1,18 +1,0 @@
-package com.codestates.main07.clubBoard.exception;
-
-import lombok.Getter;
-
-@Getter
-public enum ExceptionCode {
-    BOARD_NOT_FOUND(404, "Board not found"),
-    COMMENT_NOT_FOUND(404, "Comment not found");
-
-    private int status;
-
-    private String message;
-
-    ExceptionCode(int code, String message) {
-        this.status = code;
-        this.message = message;
-    }
-}
